fixed staging state issue with override property Fixed failing tests to do with SNAPSHOT versioning
