--- .circleci/config.yml
@@ -1,79 +1,0 @@
-###############################################################################
-# Anchors for common settings
-###############################################################################
-defaults: &defaults
-  steps:
-    - checkout
-    - restore_cache:
-        keys:
-          - v2-gradle-caches-{{ checksum "reckon-core/gradle/dependency-locks/runtimeClasspath.lockfile" }}-{{ checksum "reckon-gradle/gradle/dependency-locks/runtimeClasspath.lockfile" }}
-    - attach_workspace:
-        at: "."
-    - run:
-        name: Compile and test
-        command: ./gradlew clean check --continue
-    - save_cache:
-        paths:
-          - "~/.gradle/caches"
-        key: v2-gradle-caches-{{ checksum "reckon-core/gradle/dependency-locks/runtimeClasspath.lockfile" }}-{{ checksum "reckon-gradle/gradle/dependency-locks/runtimeClasspath.lockfile" }}
-    - store_test_results:
-        path: "~/gradle-test-results"
-        when: always
-branches-and-tags: &branches-and-tags
-  filters:
-    branches:
-      ignore:
-        - gh-pages
-    tags:
-      only: /^\d+\.\d+\.\d+(?:-.+)?$/
-tags-only: &tags-only
-  filters:
-    branches:
-      ignore: /.*/
-    tags:
-      only: /^\d+\.\d+\.\d+(?:-.+)?$/
-
-###############################################################################
-# Jobs
-###############################################################################
-version: 2
-jobs:
-  build-java-8:
-    <<: *defaults
-    docker:
-      - image: ajoberstar/circleci-openjdk-gradle:8-jdk
-  build-java-11:
-    <<: *defaults
-    docker:
-      - image: ajoberstar/circleci-openjdk-gradle:11-jdk
-  publish:
-    docker:
-      - image: ajoberstar/circleci-openjdk-gradle:8-jdk
-    steps:
-      - checkout
-      - restore_cache:
-          keys:
-            - v2-gradle-caches-{{ checksum "reckon-core/gradle/dependency-locks/runtimeClasspath.lockfile" }}-{{ checksum "reckon-gradle/gradle/dependency-locks/runtimeClasspath.lockfile" }}
-      - run:
-          name: Decrypt Gradle properties
-          command: openssl enc -d -aes256 -base64 -in gradle.properties.enc -out gradle.properties -pass 'pass:$GRADLE_PROPS_KEY'
-      - run:
-          name: Publish Artifacts
-          command: ./gradlew publish publishPlugins
-###############################################################################
-# Workflows
-###############################################################################
-workflows:
-  version: 2
-  main:
-    jobs:
-      - build-java-8:
-          <<: *branches-and-tags
-      - build-java-11:
-          <<: *branches-and-tags
-      - publish:
-          <<: *tags-only
-          requires:
-            - build-java-8
-            - build-java-11
-          context: ajoberstar
