--- gradle.properties.enc
@@ -1,3 +1,0 @@
-U2FsdGVkX1+AN/fTf3TONhKqOeBIZQaIkBabB6NGrzQdSAQ1V3LaUw/QgUfx5ov8
-ym+Bfc2tQQsyTzVJOK4aY6iOsL3wQ+QZnRUafu+kWr1z+Wqq8oQfWcx/QSLU+Jyy
-mAahBl0QD2Z0b38xebOakVky/F4/g1NFrLLn9pBC9BA=
