--- gradle/bintray.gradle
@@ -1,16 +1,0 @@
-allprojects {
-  plugins.withId('maven-publish') {
-    publishing {
-      repositories {
-        maven {
-          name = 'bintray'
-          url = 'https://api.bintray.com/maven/ajoberstar/maven/reckon/;publish=1'
-          credentials {
-            username = System.env['BINTRAY_USER']
-            password = System.env['BINTRAY_KEY']
-          }
-        }
-      }
-    }
-  }
-}
