--- src/main/java/org/openmicroscopy/reckon/core/VcsInventorySupplier.java
@@ -1,9 +1,0 @@
-package org.openmicroscopy.reckon.core;
-
-/**
- * This is intentionally package private.
- */
-@FunctionalInterface
-interface VcsInventorySupplier {
-    VcsInventory getInventory();
-}
