--- src/main/java/org/openmicroscopy/reckon/gradle/SnapshotOptions.java
@@ -1,6 +1,0 @@
-package org.openmicroscopy.reckon.gradle;
-
-public interface SnapshotOptions {
-
-
-}
