+++ src/main/java/org/openmicroscopy/release/core/VcsInventorySupplier.java
@@ -1,0 +1,9 @@
+package org.openmicroscopy.release.core;
+
+/**
+ * This is intentionally package private.
+ */
+@FunctionalInterface
+interface VcsInventorySupplier {
+    VcsInventory getInventory();
+}
