+++ src/main/java/org/openmicroscopy/release/gradle/SnapshotOptions.java
@@ -1,0 +1,6 @@
+package org.openmicroscopy.release.gradle;
+
+public interface SnapshotOptions {
+
+
+}
