+++ src/test/resources/simplelogger.properties
@@ -1,1 +1,1 @@
-org.slf4j.simpleLogger.log.org.openmicroscopy.reckon=DEBUG
+org.slf4j.simpleLogger.log.org.openmicroscopy.release=DEBUG
